arm64: morello: Signal handling support
Since we save and restore additional registers for Morello
(capability GPRs, CSP, PCC) during kernel_entry/kernel_exit, we also
need to save this context to the user stack when a signal handler is
invoked and restore it when it returns. This is done by adding a new
morello_context struct in the signal frame and saving/restoring the
Morello registers in pt_regs there.

Like for the core registers, this allows a signal handler to modify the
interrupted context by writing to the saved Morello registers (the
uapi header uses the CHERI type __uintcap_t if available to
facilitate this). However, the usual rules regarding the merging of
64-bit registers into capability registers apply, which means that a
signal handler will need to modify both the capability view (e.g.
cregs[0]) and the 64-bit view (e.g. regs[0]) to set the desired
capability register in the interrupted context. This has no impact
for processes unaware of capabilities.

Registers stored in thread_struct (morello_user_state) are not saved
to the signal frame, as they do not need to be modified to invoke
the signal handler and the signal handler can write to them directly
if needed (exactly in the same way as tpidr_el0).

This commit also removes PSR_C64_BIT from SPSR_EL1_AARCH64_RES0_BITS
to allow returning to a context running in C64.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 33153f94f263744af1baffbd1208f0cdc865f7a6)
[ Christian A. Ehrhardt: Conflict resolution ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>