arm64: morello: ptrace: Add read-only regset for Morello registers
This patch adds a new regset (NT_ARM_MORELLO) allowing all user
Morello registers to be read into struct user_morello_state, using:

  ptrace(PTRACE_GETREGSET, pid, NT_ARM_MORELLO, &iov);

Capability registers are not stored in the struct as (tagged)
capabilities, as this would create capabilities tied to an unrelated
address space in the tracer. Instead, for each capability register,
the untagged 128-bit value is stored in the corresponding field, and
its tag is stored separately in the tag_map field.

No setter is provided for this regset, because this would allow the
tracer to create arbitrary valid capabilities in the tracee,
including capabilities that the tracee could not otherwise obtain.
A new ptrace operation for manipulating the tracee's capability
registers in a safe way will be provided at a later stage.

It remains however possible to set the standard 64-bit registers via
the existing NT_PRSTATUS regset. Doing so will be handled as usual
by the kernel, that is the new 64-bit value will be merged into the
corresponding capability register the next time it is read,
potentially invalidating the capability.

Since part of the Morello registers are stored in thread_struct, we
need to flush them before accessing thread_struct if the target is
current, for the same reason as in tls_get().

Note: the new regset will be dumped in core files, like the other
regsets.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit a59bc3af1c70a3012fb9a08bf7dcd976c0cbecfb)
[ Christian A. Ehrhardt:
  - Standard conflict resolution
  - Value for NT_ARM_MORELLO changes. ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>