arm64: morello: ptrace: Add request for reading a capability in memory
This patch introduces a new arch-specific PTRACE_PEEKCAP request
type, allowing the tracer to read a capability from the tracee's
memory, using:

  ptrace(PTRACE_PEEKCAP, pid, addr, &user_cap);

addr must be capability-aligned (16-byte alignment). user_cap is a
struct containing the 128-bit value and separately the tag of the
capability read from the tracee, on the same principle as the
NT_ARM_MORELLO regset.

Since ptrace_access_vm() ultimately uses copy_from_user_page() to
copy data from the tracee's memory, we cannot use it as-is to
implement PTRACE_PEEKCAP, because copy_from_user_page() does not
preserve capability tags. While we could change
copy_from_user_page() to always copy tags, this may lead to
capabilities being needlessly created in the kernel, and would be
inconsistent with the fine-grained approach we are taking so far
regarding the manipulation of capabilities in the kernel. Instead,
we are adding a simple primitive to access a capability in another
task, loosely based on ptrace_access_vm().

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit b8f41fbfe361113e395ca25a591dd5b008abb789)
[ Christian A. Ehrhardt: Conflict resolution ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>