arm64: morello: Override user_ptr conversion helpers for PCuABI
When targeting the pure-capability uABI, implement the
*uaddr_to_user_ptr* helpers by deriving a user pointer from the
appropriate root capability.

Including <cheriintrin.h> in user_ptr.h would result in a lot of .c
files including it indirectly, which is not necessarily desirable.
To avoid this, CHERI builtins are used directly, as an exception.

Note: we are not currently using the appropriate root capability for
userspace addresses, this should be changed once the user root
capability is introduced.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 97e747b076d27e2c6c8fd5f72b7281dde218a245)
[ Christian A. Ehrhardt: Conflict resolution ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>