arm64: uaccess: Support capabilities in __uaccess_mask_ptr()
__uaccess_mask_ptr() manipulates user pointers in inline assembly,
so we need a different implementation when targeting PCuABI.

Since 2305b809be93 ("arm64: uaccess: simplify uaccess_mask_ptr()"),
invalid (non-capability) pointers are handled simply by clearing bit
55. Unfortunately this approach is probably not ideal in PCuABI, as
this would systematically require using an SCVALUE instruction to
update the address of the capability.

Keeping a similar approach to before this commit for PCuABI
therefore seems preferable, especially since CSEL is able to handle
capabilities directly: if bit 55 is set, return the null capability,
otherwise return the input capability unchanged.

Note that Morello Clang does not allow to refer to a capability
register as an X register at the time of writing, so we need to
allocate another register to represent the address part of the input
capability. This is not a huge concern though as
__uaccess_mask_ptr() is typically paired with access_ok(), which
also inspects the capability's address.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 222df10724a397d1d5463ee0472d9954ebe9f5d3)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>