arm64: morello: Disable trapping early and unconditionally
We are about to turn the kernel into a hybrid-capability binary by
allowing capabilities to be used directly in C. As a result, Morello
instructions may be used in any C function, which means that we need
to enable Morello instructions (i.e. stop trapping them) very early
in the boot sequence, before any C function is called.

We can therefore no longer wait to reach the CPU feature framework
to initialise the Morello state; this needs to happen during the
early (primary/secondary) CPU initialisation. __cpu_setup() is
the obvious place to do this.

Another complication is created by the hierarchical trapping
mechanism, requiring Morello trapping to be disabled at both EL2 and
EL1 to be able to use Morello instructions at EL1. Since commit
0c93df9622d4 ("arm64: Initialise as nVHE before switching to VHE"),
the kernel always drops to EL1 if started at EL2, and only then
chooses whether to enable VHE and return to EL2. Crucially, C
functions are called from __primary_switched() before VHE is
potentially enabled. This means that even if we assume that VHE is
used, we must setup Morello at EL2 explicitly to allow it to be used
at EL1. This also has the side-effect of enabling capability
exception entry/return at EL2 right away, requiring us to set
CVBAR_EL2 and CELR_EL2 to appropriate capabilities.

Note that if the kernel is started at EL1, there is no need to
disable Morello trapping before __cpu_setup() is reached (so
init_el1() is unaffected).

An immediate consequence of allowing Morello to be used directly in
C is that any attempt at preserving runtime detection of Morello is
futile, since Morello instructions may be used anywhere and
unconditionally. In other words, we have to throw in the towel and
give up on the single kernel image: a Morello kernel image can now
only boot on Morello hardware. While it would be theoretically
possible to error out if Morello support is enabled but the hardware
does not support it, this is very hard to do in practice (because no
C code can be safely called in that case), and it is unlikely that
anyone would attempt to boot such a kernel on non-Morello hardware
anyway.

This patch therefore removes all runtime checks based on ID
registers in the early startup code, making Morello support a pure
compile-time decision (based on CONFIG_ARM64_MORELLO). This also
allows us to simplify a few things, for instance modifying directly
the places that set VBAR_ELx, instead of using a read-modify-write
sequence later to initialise CVBAR_ELx.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit ac6ffa8a1fe8b371b143388928f3eb3c99f77408)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>