kernel/ptrace: Support compat64 in compat_ptrace_request()
compat_ptrace_request() does not currently handle all requests where
the addr or data argument is a user pointer. Make it do so in order
to support compat64, in which case compat_ptr() must always be used
to obtain a valid user pointer from a compat user pointer.

To avoid duplication, the handling of PTRACE_{GET,SET}SIGMASK is
moved to helpers that are now called from both ptrace_request() and
compat_ptrace_request() (like most other requests).

Note that PTRACE_PEEKSIGINFO is a special case as it interprets both
addr and data as user pointers, but ptrace_peek_siginfo() converts
the data argument to a user pointer itself if currently in a compat
syscall, so there is no need to do so in compat_ptrace_request().

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 72eafde8f4eda724f9587a44398059b21ab3ea6f)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>