usb: Annotate user pointers with __user
The definition of USBDEVFS_REAPURB{,NDELAY} is currently based on
the size of void *, which is not appropriate in PCuABI where user
pointers are of a different size. Annotate the pointers with __user
to fix this.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 40249df8e47f2ddb6c53d984d48d5b81a378934a)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>