kselftests/arm64/morello: Pass the positive exit() tests
The current exit syscall tests pass numbers to denote the exit status
which have to be interpreted in a special way to pass or fail the test.
Keep the exit status as EXIT_SUCCESS (0) so as to treat the successful
termination in a generic way.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
(cherry picked from commit c4c70a9e2dc629596eb02e15e2e535247c493c71)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>