fork: clone: Pass through newsp/tls clone args unmodified in compat
A subtle ABI change was introduced in compat64 by "clone: Alter
clone to accept capabilities". Indeed, since there is no compat
handler for clone, the native one is also used for compat, including
64-bit compat (compat64). This is where the way we convert compat64
syscall arguments to native (PCuABI) is showing its limits: if a
syscall wrapper expects a capability-sized argument, compat_ptr() is
used to convert the user-provided 64-bit value to a capability.

In general, this is what we want, and in fact it is the case for the
parent_tidptr and child_tidptr arguments of clone, which are
ordinary pointers to user data. However, the newsp and tls arguments
are special: they specify the value to set registers to. We should
not alter these values in any way: in arm64/PCuABI, they are
capabilities and we set CSP/CTPIDR accordingly, but in hybrid, they
are still 64-bit values and we should only set the lower 64 bits of
CSP/CTPIDR. This is not the case in compat64 as compat_ptr() is
called to turn these 64-bit values into capabilities.

The most correct solution would be to introduce a compat clone
wrapper, but this is rather painful as clone has 4 possible
prototypes depending on the architecture. The approach taken here is
a middle ground, narrowing down the stack / TLS pointer arguments
in the native handler if we got called from compat. This effectively
cancels out the automatic creation of capabilities in the compat
syscall wrapper, which is not ideal but considered acceptable in
this very particular situation.

Fixes: ("clone: Alter clone to accept capabilities")
Co-developed-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 887989a248ba7461bac81039b4ac2a13a6b86de9)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>