kselftests/arm64: morello: Fix restricted mode tests with Gcc
Some of the Gcc compiled clone tests in restricted mode face stack
corruption. This occurs as the run-time permission settings (executive or
restricted mode) in the PCC may be lost if function pointers are
materialised from the GOT entries. This further leads to selecting
an incorrect stack pointer(CSP instead of RCSP) and hence the crash.

As a workaround, materialise the function pointer clone_base_fn explicitly
by building branch address from the PCC. Here it is assumed that the bounds
and permissions of clone_base_fn are same as the PCC , which is not an
issue in practice (being in the same translation unit).

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
(cherry picked from commit 03e83b6b188e1359c747f9f759fa2ee679448d93)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>