bpf: use addr for bpf_copy_from_user_with_task
The purpose of the bpf_copy_from_user_with_task eBPF helper is to copy
user data from the provided task. Since __user *ptrs are only valid for
the current task/process, what is actually being passed in here is an
address. access_process_vm is then used to access that address given the
task specified by the task_struct.

Additionally, this helper is called from an eBPF program, which does not
support capabilities. It is therefore unable to pass in a capability as
a __user ptr in PCuABI in the first place.

The use case of this helper is to read arbitrary user memory, as used by
security or monitoring eBPF programs. There is therefore no requirement
to check user capabilities here. Loading of eBPF applications using this
helper type is already strictly limited to privileged/root users.

In addition, access_process_vm() is not uaccess in the usual sense, as
uaccess is about accessing the current process. This type of access is
not currently checked in PCuABI.

Change the inbound type from __user * to ptraddr_t and remove the TODO
since no capability checks are required here.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
(cherry picked from commit 0d04fb14f64c6054c6218949ccb9e37ba6b5e108)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>