bpf: use addr for bpf_copy_from_user
The purpose of this eBPF helper function is to read user memory, as used
by security or monitoring eBPF programs. There is therefore no
requirement to check user capabilities or uaccess here. Loading of eBPF
applications using this helper type is already strictly limited to
privileged/root users.

Additionally, this helper is called from an eBPF program, which does not
support capabilities. It is therefore unable to pass in a capability as
a __user ptr in PCuABI in the first place, so what is actually passed in
is an address.

In order to access user memory with just an address, use
make_user_ptr_for_read_uaccess() to generate a capability of appropriate
bounds for the kernel to use.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
(cherry picked from commit ec41a8edd9ed454e80f1eac0ea21b6d626b46092)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>