arm64: morello: Zero out C reg in pt_regs_write_reg()
As per the register merging principle set out in the Morello
documentation [1], the kernel does not normally zero out the C
register when writing to the corresponding X register. However,
pt_regs_write_ptr_reg() is used in contexts such as instruction
emulation, where the kernel emulates the effect of an instruction
writing to an X register.

Instructions writing to X registers are architecturally guaranteed
to clear the C register (that is the upper 64 bits and tag); it is
therefore preferable for pt_regs_write_reg() to behave in the same
way.

[1] Documentation/arch/arm64/morello.rst

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit a0508cc5a72d122acda1bf46896676acbb95b355)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>