kselftests/arm64: morello: Remove invalid assumptions about initial data
The bootstrap tests make a number of assumptions that are not in
line with the PCuABI specification, notably that argv/envp have the
same bounds as CSP and that AT_ENTRY is unsealed. We are about to
improve the alignment with the spec, so remove those assumptions.

Since the initial data cannot be assumed to live on the stack at all
in PCuABI, the parsing of argc/argv/envp/auxv on the stack is
removed, and instead the argv/envp pointers (passed in c1/c2) are
asserted to be equal to AT_ARGV/AT_ENVP.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit e6391ac1a93f6c6d718d649a78bdb01beb2c291e)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>