wext: use packed lengths in packed stream
Update the lengths that are returned by iwe_stream_lcp_len and
iwe_stream_point_len.

These lengths are used for storing data fields in a packed stream. We have
to return the lengths of the corresponding packed fields, not the lengths
that include padding.

Change the lengths only for cheri kernels. For 64bit non-cheri systems,
there are already workarounds in wireless-tools to remove the padding.
(For 32 bit, there is no padding.)