riscv: ucontext.h: Fix build error with libc definition of sigset_t
Building a user space program which includes <asm/ucontext.h> with
CHERI enabled leads to a build error:
| ../recipe-sysroot/usr/include/asm/ucontext.h:39:19: error: array is too large (18446744073709551600 elements)
|    39 |         __u8              __unused[1024 / 8 - sizeof(sigset_t) - 2 * __SIZEOF_LONG__];
|       |                                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Before the introduction of the CHERI changes, the purpose of this
__unused array was simply to add padding, to ensure that a sigset_t
plus the padding was always 1024 bits long, irrespective of the size of
sigset_t. If sigset_t was already 1024 bits, which it is when using
the musl or glibc definition, __unused was of zero size.

However, with the introduction of CHERI two longs worth of the padding
have been removed, so the combined size of sigset_t and __unused must
be no more than 896 bits. When sigset_t is already 1024 bits long this
calculation results in a negative array size, hence this error.

In practice, only the first 64 bits of a sigset_t are ever used on
RISC-V, so introduce a __kernel_sigset_t, which is the correct size,
and use this in the definition of struct ucontext.