trace: cheri: fix temporary event for filtering
When a filter is set, logged events are stored in a temporary ringbuffer
event before the filter is applied.

For such a temporary event, populate the length field in array[0] before
we use it for the bounds of the data capability.

Fix the length field calculation, the length field includes the size of
array[0].