riscv: dts: codasip: Prime platform
This is a platform to enable evaluation of CHERI, particularly CHERI
Linux. It is based on the standard VCU118 platform, with the addition of
PCIe express. The PCIe express allows the connection of NVME drives,
which is a massive improvement over the very slow SD card that has been
the only storage up to now.

The prime platform uses the "Lux" core. This config has 4 cores, 2M of
L2 cache, and has the Zbkc extension to improve coremark.

The deprecated riscv,isa property is retained for compatibility with
operating systems such as BSD which does not support the newer
property.

Signed-off-by: David McKay <david.mckay@codasip.com>