cheri: io_uring: Use type fitting 64-bit ABI members
Replace the type user_uintptr_t with __uptr64 wherever a 64-bit
ABI field is passed from user space to the kernel, as
user_uintptr_t is not wide enough for a 64-bit value on 32-bit
non-CHERI. This is especially critical for e.g. the 64-bit SQE
user_data member, which must not be modified by the kernel, even
on 32-bit architectures.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>