cheri: lib/sort: Check base for 16 byte align swap
The sort code has been extended to use capability load/stores to swap
entries. It checks if the size is a multiple of 16 bytes to do this.

However, for CHERI it is also essential to check the alignment of the
base address as well, otherwise it may attempt to use capability
load/stores on unaligned addresses which will generate a fault. This
check was being omitted because the code assumed that the alignment of
the base address was irrelevant if we have fast unaligned support.

While there fix the swap functions for 32-bit CHERI which has 64-bit
pointers, too.

Signed-off-by: David McKay <david.mckay@codasip.com>