trace: update snapshot functionality for cheri
The tracing framework supports taking a snapshot of the ringbuffer,
either on an explicit request from the user or when a filter is hit.

Update this snapshot code to work with cheri.

The changes were tested by setting a filter such as

   echo "_printk:snapshot:4" > /sys/kernel/tracing/set_ftrace_filter

The next 4 times that _prinkt is called will trigger a snapshot of the
ringbuffer.

The counter from the filter is parsed as an unsigned value and stored in
a pointer. When a snapshot is taken and the counter is decremented, it's
treated as a signed value. We don't change this for cheri.